Filters/requirements.txt
numpy==1.24.1
opencv_python_headless==4.7.0.68
Pillow==9.4.0
scipy==1.9.3
streamlit==1.17.0
